revert

+++ orchid-backend-net/orchid-backend-net.Application/Method/Dto/Method/GetSuccessCompletedExperimentLogRateDto.cs
@@ -14,7 +14,7 @@ namespace orchid_backend_net.Application.Method.Dto.Method
         public int? TotalExperimentLog { get; set; }
         public int? CompletedExperimentLog { get; set; }
         public int? FailedExperimentLog { get; set; }
-        public double? SuccessRate { get; set; }
+        public int? SuccessRate { get; set; }
         public List<Seedlings>? Seedling { get; set; }
         public List<Domain.Entities.MethodStageDefinition>? MethodStages { get; set; }
     }
